requirements.txt
# Library requirements for Setting Up the Environment for the PDF Reader Code
PyPDF2==3.0.1
pytesseract==0.3.15
Pillow==11.1.0
numpy==2.1.3
PyMuPDF==1.26.1
tabula-py==2.10.0
pdf2image==1.17.0
pandas==2.2.3

